fix: fixed old issues improved code deleted unused resources
