+++ src/constants/index.ts
@@ -1,6 +1,7 @@
 export const ERRORS = {
   support: `Your browser doesn't support Intersection Observer API, try update browser or use another one. Check if your browser support Intersection Observer API here: https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API`,
 };
+
 export const STYLES = {
   defaultRootMargin: '0px 0px 50px 0px',
   fallBackDefault: {
@@ -10,6 +11,7 @@ export const STYLES = {
     top: '35%',
   },
 };
+
 export const CLASS_NAMES = {
   transparent_default: 'transparent-default',
   backdrop_default: 'backdrop-default',
