+++ src/image-worker/interfaces/types.ts
@@ -1,8 +1,5 @@
-import { FunctionComponent, ComponentClass } from 'react';
-
 export type stringOrArray = HTMLImageElement | string | [string];
 export type stringNumber = string | number;
-export type reactNode = string | FunctionComponent | ComponentClass | undefined;
 export type numberOrUndefined = number | undefined;
 export type stringOrUndefined = string | undefined;
 export type element = Element | null | undefined;
