+++ src/index.ts
@@ -1,3 +1,2 @@
-import { Reimage } from './image-worker';
-
+import { Reimage } from './image-worker/index';
 export default Reimage;
